refactor `favorites-page` components
