+++ src/components/header/components/logged-nav.tsx
@@ -1,7 +1,7 @@
 import { useAuth } from '../../../hooks/user-authorization';
 import { userActions, userSelector } from '../../../store/slices/user';
 import { useActionCreators, useAppSelector } from '../../../hooks/store';
-import useFavoriteCount from '../../../hooks/use-favorite-count';
+import { useFavoriteCount } from '../../../hooks/use-favorite-count';
 
 import { Link } from 'react-router-dom';
 import { AppRoute } from '../../../utils/data';
