+++ src/pages/favorites-page/favorites-page.tsx
@@ -1,22 +1,19 @@
 import { useAppSelector } from '../../hooks/store';
-import usePageLayout from '../../hooks/use-page-layout';
 
-import EmptyFavoritesSection from './components/empty-favorites-section';
-import FilledFavoritesSection from './components/filled-favorites-page';
+import { EmptyFavoritesSection } from './components/empty-favorites-section';
+import { FilledFavoritesSection } from './components/filled-favorites-section';
 
 import { favoritesSelector } from '../../store/slices/favorites';
 
-export default function FavoritesPage() {
+function BaseFavoritesPage() {
   const favorites = useAppSelector(favoritesSelector.favorites);
   const favoritesLength = favorites.length;
 
-  const { emptyFavorites } = usePageLayout({
-    favoritesLength,
-  });
+  const isEmpty = favoritesLength === 0;
 
   return (
     <div className="page__favorites-container container">
-      {emptyFavorites ? (
+      {isEmpty ? (
         <EmptyFavoritesSection />
       ) : (
         <FilledFavoritesSection favorites={favorites} />
@@ -24,3 +21,5 @@ export default function FavoritesPage() {
     </div>
   );
 }
+
+export const FavoritesPage = BaseFavoritesPage;
